requirements.txt
torch>=1.13.1
transformers>=4.30.0
python-dotenv>=1.0.0
